+++ src/server/config/iUnifiSettings.ts
@@ -1,0 +1,5 @@
+import { iUnifiConnectionOptions } from './iUnifiConnectionOptions';
+
+export interface iUnifiSettings {
+  loginOptions: iUnifiConnectionOptions;
+}