+++ src/server/config/index.ts
@@ -22,4 +22,6 @@ export * from './iTelegramSettings';
 export * from './iTibberSettings';
 export * from './iTimeSettings';
 export * from './iTimePair';
+export * from './iUnifiSettings';
+export * from './iUnifiConnectionOptions';
 export * from './iVictronSettings';