+++ src/server/services/network/index.ts
@@ -1,2 +1,3 @@
 export * from './asusRouter';
+export * from './unifiRouter';
 export * from './router';